+++ src/api/OrderComplete.ts
@@ -1,11 +1,11 @@
-import axios from 'axios';
+import { axiosWithAccessToken } from '../Axios';
 
 const OrderAxios = async () => {
   try {
-    const response = await axios.post('/api/orders');
-    return response.data.message;
+    const response = await axiosWithAccessToken.get('/api/reservations');
+    return response.data.data;
   } catch (error) {
-    console.log(error);
+    console.error(error);
     return error;
   }
 };
